Terrain.py issues with newer digimap tiles
Code does not combine delimiters when splitting lines into items and creates empty string values that are resolved with an if item != '', and the line is stripped prior to delimiter splitting to catch any extra spaces at the ends of each line.

Newer digimap composite tiles include an additional information line defining no_data areas with the value of -9999. this additional line breaks the code as it is 6 terms long and not the same as the ncol value for the file. 

Additionally, any -9999 values don't throw any errors but significantly throw off the visibility of minor z value deviations on the plotted surface, due to the "cliff" forming from regular values to -9999.